python-api/requirements.txt
fastapi
uvicorn
pydantic
earthengine-api
google-auth
google-auth-oauthlib
